Program to print out Whether the given element is Prime As well as Palindrome
In the code , two methods are present, one method is isPrime() checks for the given number is prime or not and the other method is isPalindrome() to check whether 
the number is palindrome or not. We have list of numbers which will be using as input for the above methods , checking the given number is prime and palindrome by using 
Filter() method.